+++ src/components/Dialog/ActionDialog/ActionDialogContentInner.tsx
@@ -138,7 +138,7 @@ export const ActionDialogContentInner: FC<ActionDialogContentInnerProps> = ({
 
 const TitleArea = styled(Stack).attrs(() => ({
   gap: 0.25,
-  as: 'span',
+  forwardedAs: 'span',
 }))<{ themes: Theme }>`
   ${({ themes: { border, space } }) => css`
     margin-block: unset;