+++ src/components/Layout/Reel/Reel.tsx
@@ -1,15 +1,16 @@
-import React, { ComponentProps, PropsWithChildren, useMemo } from 'react'
+import React, { forwardRef, useMemo } from 'react'
 import { VariantProps, tv } from 'tailwind-variants'
 
 import { useSectionWrapper } from '../../SectioningContent/useSectioningWrapper'
 
 import type { Gap } from '../../../types'
+import type { ComponentPropsWithRef, PropsWithChildren } from 'react'
 
 type Props = VariantProps<typeof reel> &
   PropsWithChildren<{
     as?: string | React.ComponentType<any>
   }> &
-  ComponentProps<'div'>
+  ComponentPropsWithRef<'div'>
 
 const reel = tv({
   base: [
@@ -85,20 +86,16 @@ const reel = tv({
   },
 })
 
-export const Reel: React.FC<Props> = ({
-  as: Component = 'div',
-  gap = 0.5,
-  padding = 0,
-  className,
-  ...props
-}) => {
-  const styles = useMemo(() => reel({ gap, padding, className }), [className, gap, padding])
+export const Reel = forwardRef<HTMLDivElement, Props>(
+  ({ as: Component = 'div', gap = 0.5, padding = 0, className, ...props }, ref) => {
+    const styles = useMemo(() => reel({ gap, padding, className }), [className, gap, padding])
 
-  const Wrapper = useSectionWrapper(Component)
+    const Wrapper = useSectionWrapper(Component)
 
-  return (
-    <Wrapper>
-      <Component {...props} className={styles} />
-    </Wrapper>
-  )
-}
+    return (
+      <Wrapper>
+        <Component {...props} ref={ref} className={styles} />
+      </Wrapper>
+    )
+  },
+)