+++ src/smarthr-ui-preset.ts
@@ -30,6 +30,7 @@ export default {
       'base-grey': defaultColor.BASE_GREY,
       'over-background': defaultColor.OVER_BACKGROUND,
       head: defaultColor.HEAD,
+      'head-darken': darkenColor(defaultColor.HEAD),
       'action-background': defaultColor.ACTION_BACKGROUND,
       main: defaultColor.MAIN,
       'main-darken': theme('colors.main-darken'),
@@ -132,6 +133,7 @@ export default {
       black: defaultColor.GREY_100,
     },
     textColor: ({ theme }) => ({
+      main: theme('colors.main'),
       black: theme('colors.black'),
       white: theme('colors.white'),
       'white-darken': theme('colors.white-darken'),
@@ -151,6 +153,9 @@ export default {
       'flash-message': `${defaultZIndex.FLASH_MESSAGE}`,
     },
     extend: {
+      minHeight: ({ theme }) => ({
+        ...theme('spacing'),
+      }),
       borderColor: ({ theme }) => ({
         default: theme('colors.grey.20'),
         disabled: theme('colors.grey.20 / 50%'),